requirements.txt
openai
flask
apscheduler[psycopg,sqlalchemy]
numpy
mysql-connector-python
faiss-cpu
gunicorn